+++ Dfe.Academies.Academisation.Data/Migrations/20260903091629_AddSigChangeEqualitiesImpactAssessment.Designer.cs
@@ -4,6 +4,7 @@ using Dfe.Academies.Academisation.Data;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.EntityFrameworkCore.Infrastructure;
 using Microsoft.EntityFrameworkCore.Metadata;
+using Microsoft.EntityFrameworkCore.Migrations;
 using Microsoft.EntityFrameworkCore.Storage.ValueConversion;
 
 #nullable disable
@@ -11,9 +12,11 @@ using Microsoft.EntityFrameworkCore.Storage.ValueConversion;
 namespace Dfe.Academies.Academisation.Data.Migrations
 {
     [DbContext(typeof(AcademisationContext))]
-    partial class AcademisationContextModelSnapshot : ModelSnapshot
+    [Migration("20260903091629_AddSigChangeEqualitiesImpactAssessment")]
+    partial class AddSigChangeEqualitiesImpactAssessment
     {
-        protected override void BuildModel(ModelBuilder modelBuilder)
+        /// <inheritdoc />
+        protected override void BuildTargetModel(ModelBuilder modelBuilder)
         {
 #pragma warning disable 612, 618
             modelBuilder
@@ -886,19 +889,11 @@ namespace Dfe.Academies.Academisation.Data.Migrations
                         .HasColumnType("uniqueidentifier")
                         .HasColumnName("AssignedUserId");
 
-                    b.Property<string>("CompaniesHouseNumber")
-                        .HasColumnType("nvarchar(max)")
-                        .HasColumnName("CompaniesHouseNumber");
-
                     b.Property<DateTime>("CreatedOn")
                         .HasColumnType("datetime2");
 
                     b.Property<DateTime>("LastModifiedOn")
                         .HasColumnType("datetime2");
-
-                    b.Property<string>("LocalAuthorityName")
-                        .HasColumnType("nvarchar(max)")
-                        .HasColumnName("LocalAuthorityName");
 
                     b.Property<DateTime?>("ReadOnlyDate")
                         .HasColumnType("datetime2");
@@ -2472,21 +2467,6 @@ namespace Dfe.Academies.Academisation.Data.Migrations
                             b1.Property<string>("EqualitiesImpactIdentifiedMitigation")
                                 .HasColumnType("nvarchar(max)")
                                 .HasColumnName("EqualitiesImpactIdentifiedMitigation");
-                            b1.Property<DateTime?>("ProposedChangeDate")
-                                .HasColumnType("datetime2")
-                                .HasColumnName("ProposedChangeDate");
-
-                            b1.Property<DateTime?>("ProposedDecisionDate")
-                                .HasColumnType("datetime2")
-                                .HasColumnName("ProposedDecisionDate");
-
-                            b1.Property<bool?>("TrustConsultedReligiousBody")
-                                .HasColumnType("bit")
-                                .HasColumnName("TrustConsultedReligiousBody");
-
-                            b1.Property<string>("TrustConsultedReligiousBodyNotConsultedReason")
-                                .HasColumnType("nvarchar(max)")
-                                .HasColumnName("TrustConsultedReligiousBodyNotConsultedReason");
 
                             b1.Property<bool?>("TrustConsultedStakeholders")
                                 .HasColumnType("bit")