+++ Dfe.Academies.Academisation.Domain.UnitTest/SignificantChange/SignificantChangeProjectTests.cs
@@ -14,14 +14,28 @@ namespace Dfe.Academies.Academisation.Domain.UnitTest.SignificantChange
 		public void Constructor_ShouldSetPropertiesCorrectly()
 		{
 			var status = _fixture.Create<SignificantChangeStatus>();
+
 			var urn = _fixture.Create<int>();
 			var tier = _fixture.Create<byte>();
 			var trustName = _fixture.Create<string>();
 			var trustUkprn = _fixture.Create<string>();
 			var typeOfSignificantChange = _fixture.Create<string>();
-			var schoolName= _fixture.Create<string>();
+			var schoolName = _fixture.Create<string>();
+			var localAuthorityName = _fixture.Create<string>();
+			var companiesHouseNumber = _fixture.Create<string>();
 
-			var project = new SignificantChangeProject(status, urn, tier, trustName, trustUkprn, typeOfSignificantChange, schoolName);
+			var significantChangeProjectOptions = new SignificantChangeProjectOptions(
+				urn,
+				tier,
+				trustName,
+				trustUkprn,
+				typeOfSignificantChange,
+				schoolName,
+				localAuthorityName,
+				companiesHouseNumber
+			);
+
+			var project = new SignificantChangeProject(status, significantChangeProjectOptions);
 
 			project.Status.Should().Be(status);
 			project.Urn.Should().Be(urn);
@@ -29,6 +43,9 @@ namespace Dfe.Academies.Academisation.Domain.UnitTest.SignificantChange
 			project.TrustName.Should().Be(trustName);
 			project.TrustUkprn.Should().Be(trustUkprn);
 			project.TypeOfSignificantChange.Should().Be(typeOfSignificantChange);
+			project.SchoolName.Should().Be(schoolName);
+			project.LocalAuthorityName.Should().Be(localAuthorityName);
+			project.CompaniesHouseNumber.Should().Be(companiesHouseNumber);
 			project.AssignedUserId.Should().BeNull();
 			project.AssignedUserFullName.Should().BeNull();
 			project.AssignedUserEmailAddress.Should().BeNull();
@@ -37,8 +54,9 @@ namespace Dfe.Academies.Academisation.Domain.UnitTest.SignificantChange
 		[Fact]
 		public void AssignUser_ShouldSetUserProperties()
 		{
-			var project = new SignificantChangeProject(
-				_fixture.Create<SignificantChangeStatus>(),
+			var status = _fixture.Create<SignificantChangeStatus>();
+
+			var significantChangeProjectOptions = new SignificantChangeProjectOptions(
 				_fixture.Create<int>(),
 				_fixture.Create<byte>(),
 				_fixture.Create<string>(),
@@ -46,6 +64,8 @@ namespace Dfe.Academies.Academisation.Domain.UnitTest.SignificantChange
 				_fixture.Create<string>(),
 				_fixture.Create<string>()
 			);
+
+			var project = new SignificantChangeProject(status, significantChangeProjectOptions);
 
 			var userId = _fixture.Create<Guid>();
 			var userEmail = _fixture.Create<string>();
@@ -56,19 +76,22 @@ namespace Dfe.Academies.Academisation.Domain.UnitTest.SignificantChange
 			project.AssignedUserId.Should().Be(userId);
 			project.AssignedUserEmailAddress.Should().Be(userEmail);
 			project.AssignedUserFullName.Should().Be(userFullName);
+
+
 		}
-    
+
 		[Fact]
 		public void SetReadOnlyDate_ShouldSetReadOnlyDate()
 		{
-			var project = SignificantChangeProject.Create(
+			var significantChangeProjectOptions = new SignificantChangeProjectOptions(
 				_fixture.Create<int>(),
 				_fixture.Create<byte>(),
 				_fixture.Create<string>(),
 				_fixture.Create<string>(),
 				_fixture.Create<string>(),
-				_fixture.Create<string>(),
-				DateTime.UtcNow);
+				_fixture.Create<string>()
+			);
+			var project = SignificantChangeProject.Create(significantChangeProjectOptions, DateTime.UtcNow);
 
 			var readOnlyDate = DateTime.UtcNow.AddDays(-1);
 
@@ -80,8 +103,9 @@ namespace Dfe.Academies.Academisation.Domain.UnitTest.SignificantChange
 		[Fact]
 		public void SetStakeholderConsultation_ShouldSetDetailsProperties()
 		{
-			var project = new SignificantChangeProject(
-				_fixture.Create<SignificantChangeStatus>(),
+			var status = _fixture.Create<SignificantChangeStatus>();
+
+			var significantChangeProjectOptions = new SignificantChangeProjectOptions(
 				_fixture.Create<int>(),
 				_fixture.Create<byte>(),
 				_fixture.Create<string>(),
@@ -90,17 +114,22 @@ namespace Dfe.Academies.Academisation.Domain.UnitTest.SignificantChange
 				_fixture.Create<string>()
 			);
 
+			var project = new SignificantChangeProject(status, significantChangeProjectOptions);
+
 			project.SetStakeholderConsultation(false, "Trust has not consulted stakeholders yet");
 
 			project.Details.TrustConsultedStakeholders.Should().BeFalse();
-			project.Details.TrustConsultedStakeholdersNotConsultedReason.Should().Be("Trust has not consulted stakeholders yet");
+			project.Details.TrustConsultedStakeholdersNotConsultedReason.Should()
+				.Be("Trust has not consulted stakeholders yet");
 		}
 
 		[Fact]
 		public void GetStakeholderConsultationTaskStatus_WhenNoValues_ReturnsNotStarted()
 		{
-			var project = new SignificantChangeProject(
-				_fixture.Create<SignificantChangeStatus>(),
+
+			var status = _fixture.Create<SignificantChangeStatus>();
+
+			var significantChangeProjectOptions = new SignificantChangeProjectOptions(
 				_fixture.Create<int>(),
 				_fixture.Create<byte>(),
 				_fixture.Create<string>(),
@@ -108,6 +137,8 @@ namespace Dfe.Academies.Academisation.Domain.UnitTest.SignificantChange
 				_fixture.Create<string>(),
 				_fixture.Create<string>()
 			);
+
+			var project = new SignificantChangeProject(status, significantChangeProjectOptions);
 
 			project.Details.GetStakeholderConsultationTaskStatus().Should().Be(SignificantChangeTaskStatus.NotStarted);
 		}
@@ -115,8 +146,9 @@ namespace Dfe.Academies.Academisation.Domain.UnitTest.SignificantChange
 		[Fact]
 		public void GetStakeholderConsultationTaskStatus_WhenNotConsultedWithoutReason_ReturnsInProgress()
 		{
-			var project = new SignificantChangeProject(
-				_fixture.Create<SignificantChangeStatus>(),
+			var status = _fixture.Create<SignificantChangeStatus>();
+
+			var significantChangeProjectOptions = new SignificantChangeProjectOptions(
 				_fixture.Create<int>(),
 				_fixture.Create<byte>(),
 				_fixture.Create<string>(),
@@ -124,6 +156,8 @@ namespace Dfe.Academies.Academisation.Domain.UnitTest.SignificantChange
 				_fixture.Create<string>(),
 				_fixture.Create<string>()
 			);
+
+			var project = new SignificantChangeProject(status, significantChangeProjectOptions);
 
 			project.SetStakeholderConsultation(false, null);
 
@@ -133,8 +167,8 @@ namespace Dfe.Academies.Academisation.Domain.UnitTest.SignificantChange
 		[Fact]
 		public void GetStakeholderConsultationTaskStatus_WhenConsulted_ReturnsCompleted()
 		{
-			var project = new SignificantChangeProject(
-				_fixture.Create<SignificantChangeStatus>(),
+			var status = _fixture.Create<SignificantChangeStatus>();
+			var significantChangeProjectOptions = new SignificantChangeProjectOptions(
 				_fixture.Create<int>(),
 				_fixture.Create<byte>(),
 				_fixture.Create<string>(),
@@ -142,6 +176,8 @@ namespace Dfe.Academies.Academisation.Domain.UnitTest.SignificantChange
 				_fixture.Create<string>(),
 				_fixture.Create<string>()
 			);
+
+			var project = new SignificantChangeProject(status, significantChangeProjectOptions);
 
 			project.SetStakeholderConsultation(true, null);
 
@@ -151,8 +187,9 @@ namespace Dfe.Academies.Academisation.Domain.UnitTest.SignificantChange
 		[Fact]
 		public void GetStakeholderConsultationTaskStatus_WhenNotConsultedWithReason_ReturnsCompleted()
 		{
-			var project = new SignificantChangeProject(
-				_fixture.Create<SignificantChangeStatus>(),
+			var status = _fixture.Create<SignificantChangeStatus>();
+
+			var significantChangeProjectOptions = new SignificantChangeProjectOptions(
 				_fixture.Create<int>(),
 				_fixture.Create<byte>(),
 				_fixture.Create<string>(),
@@ -160,6 +197,8 @@ namespace Dfe.Academies.Academisation.Domain.UnitTest.SignificantChange
 				_fixture.Create<string>(),
 				_fixture.Create<string>()
 			);
+
+			var project = new SignificantChangeProject(status, significantChangeProjectOptions);
 
 			project.SetStakeholderConsultation(false, "Consultation timeline does not allow this yet");
 
@@ -169,14 +208,17 @@ namespace Dfe.Academies.Academisation.Domain.UnitTest.SignificantChange
 		[Fact]
 		public void SetStakeholderConsultation_WhenNotConsulted_AndTierOne_MovesToTierTwo()
 		{
-			var project = new SignificantChangeProject(
-				SignificantChangeStatus.PreDecision,
+			var significantChangeProjectOptions = new SignificantChangeProjectOptions(
 				_fixture.Create<int>(),
 				(byte)1,
 				_fixture.Create<string>(),
 				_fixture.Create<string>(),
 				_fixture.Create<string>(),
-				_fixture.Create<string>());
+				_fixture.Create<string>()
+			);
+
+			var project =
+				new SignificantChangeProject(SignificantChangeStatus.PreDecision, significantChangeProjectOptions);
 
 			project.SetStakeholderConsultation(false, "No consultation carried out");
 
@@ -186,14 +228,18 @@ namespace Dfe.Academies.Academisation.Domain.UnitTest.SignificantChange
 		[Fact]
 		public void SetStakeholderConsultation_WhenTierMovedToTwo_DoesNotRevertToTierOne()
 		{
-			var project = new SignificantChangeProject(
-				SignificantChangeStatus.PreDecision,
+			var status = _fixture.Create<SignificantChangeStatus>();
+
+			var significantChangeProjectOptions = new SignificantChangeProjectOptions(
 				_fixture.Create<int>(),
 				(byte)1,
 				_fixture.Create<string>(),
 				_fixture.Create<string>(),
 				_fixture.Create<string>(),
-				_fixture.Create<string>());
+				_fixture.Create<string>()
+			);
+
+			var project = new SignificantChangeProject(status, significantChangeProjectOptions);
 
 			project.SetStakeholderConsultation(false, "No consultation carried out");
 			project.SetStakeholderConsultation(true, null);
@@ -202,15 +248,147 @@ namespace Dfe.Academies.Academisation.Domain.UnitTest.SignificantChange
 		}
 
 		[Fact]
-		public void SetProjectDates_ShouldSetDates()
+		public void SetReligiousBodyConsultation_ShouldSetDetailsProperties()
 		{
-			var project = SignificantChangeProject.Create(
+			var significantChangeProjectOptions = new SignificantChangeProjectOptions(
 				_fixture.Create<int>(),
 				_fixture.Create<byte>(),
 				_fixture.Create<string>(),
 				_fixture.Create<string>(),
 				_fixture.Create<string>(),
+				_fixture.Create<string>()
+			);
+
+			var project = new SignificantChangeProject(_fixture.Create<SignificantChangeStatus>(), significantChangeProjectOptions);
+
+			project.SetReligiousBodyConsultation(false, "Trust has not consulted religious body yet");
+
+			project.Details.TrustConsultedReligiousBody.Should().BeFalse();
+			project.Details.TrustConsultedReligiousBodyNotConsultedReason.Should().Be("Trust has not consulted religious body yet");
+		}
+
+		[Fact]
+		public void GetReligiousBodyConsultationTaskStatus_WhenNoValues_ReturnsNotStarted()
+		{
+			var significantChangeProjectOptions = new SignificantChangeProjectOptions(
+				_fixture.Create<int>(),
+				_fixture.Create<byte>(),
 				_fixture.Create<string>(),
+				_fixture.Create<string>(),
+				_fixture.Create<string>(),
+				_fixture.Create<string>()
+			);
+
+			var project = new SignificantChangeProject(_fixture.Create<SignificantChangeStatus>(), significantChangeProjectOptions);
+
+			project.Details.GetReligiousBodyConsultationTaskStatus().Should().Be(SignificantChangeTaskStatus.NotStarted);
+		}
+
+		[Fact]
+		public void GetReligiousBodyConsultationTaskStatus_WhenNotConsultedWithoutReason_ReturnsInProgress()
+		{
+			var significantChangeProjectOptions = new SignificantChangeProjectOptions(
+				_fixture.Create<int>(),
+				_fixture.Create<byte>(),
+				_fixture.Create<string>(),
+				_fixture.Create<string>(),
+				_fixture.Create<string>(),
+				_fixture.Create<string>()
+			);
+
+			var project = new SignificantChangeProject(_fixture.Create<SignificantChangeStatus>(), significantChangeProjectOptions);
+
+			project.SetReligiousBodyConsultation(false, null);
+
+			project.Details.GetReligiousBodyConsultationTaskStatus().Should().Be(SignificantChangeTaskStatus.InProgress);
+		}
+
+		[Fact]
+		public void GetReligiousBodyConsultationTaskStatus_WhenConsulted_ReturnsCompleted()
+		{
+			var significantChangeProjectOptions = new SignificantChangeProjectOptions(
+				_fixture.Create<int>(),
+				_fixture.Create<byte>(),
+				_fixture.Create<string>(),
+				_fixture.Create<string>(),
+				_fixture.Create<string>(),
+				_fixture.Create<string>()
+			);
+
+			var project = new SignificantChangeProject(_fixture.Create<SignificantChangeStatus>(), significantChangeProjectOptions);
+
+			project.SetReligiousBodyConsultation(true, null);
+
+			project.Details.GetReligiousBodyConsultationTaskStatus().Should().Be(SignificantChangeTaskStatus.Completed);
+		}
+
+		[Fact]
+		public void GetReligiousBodyConsultationTaskStatus_WhenNotConsultedWithReason_ReturnsCompleted()
+		{
+			var significantChangeProjectOptions = new SignificantChangeProjectOptions(
+				_fixture.Create<int>(),
+				_fixture.Create<byte>(),
+				_fixture.Create<string>(),
+				_fixture.Create<string>(),
+				_fixture.Create<string>(),
+				_fixture.Create<string>()
+			);
+
+			var project = new SignificantChangeProject(_fixture.Create<SignificantChangeStatus>(), significantChangeProjectOptions);
+
+			project.SetReligiousBodyConsultation(false, "Consultation timeline does not allow this yet");
+
+			project.Details.GetReligiousBodyConsultationTaskStatus().Should().Be(SignificantChangeTaskStatus.Completed);
+		}
+
+		[Fact]
+		public void SetReligiousBodyConsultation_WhenNotConsulted_AndTierOne_MovesToTierTwo()
+		{
+			var significantChangeProjectOptions = new SignificantChangeProjectOptions(
+				_fixture.Create<int>(),
+				(byte)1,
+				_fixture.Create<string>(),
+				_fixture.Create<string>(),
+				_fixture.Create<string>(),
+				_fixture.Create<string>());
+
+			var project = new SignificantChangeProject(SignificantChangeStatus.PreDecision, significantChangeProjectOptions);
+
+			project.SetReligiousBodyConsultation(false, "No consultation carried out");
+
+			project.Tier.Should().Be(2);
+		}
+
+		[Fact]
+		public void SetReligiousBodyConsultation_WhenTierMovedToTwo_DoesNotRevertToTierOne()
+		{
+			var significantChangeProjectOptions = new SignificantChangeProjectOptions(
+				_fixture.Create<int>(),
+				(byte)1,
+				_fixture.Create<string>(),
+				_fixture.Create<string>(),
+				_fixture.Create<string>(),
+				_fixture.Create<string>());
+
+			var project = new SignificantChangeProject(SignificantChangeStatus.PreDecision, significantChangeProjectOptions);
+
+			project.SetReligiousBodyConsultation(false, "No consultation carried out");
+			project.SetReligiousBodyConsultation(true, null);
+
+			project.Tier.Should().Be(2);
+		}
+
+		[Fact]
+		public void SetProjectDates_ShouldSetDates()
+		{
+			var project = SignificantChangeProject.Create(
+				new SignificantChangeProjectOptions(
+					_fixture.Create<int>(),
+					_fixture.Create<byte>(),
+					_fixture.Create<string>(),
+					_fixture.Create<string>(),
+					_fixture.Create<string>(),
+					_fixture.Create<string>()),
 				DateTime.UtcNow);
 
 			var proposedDecisionDate = DateTime.UtcNow.AddDays(10);
@@ -227,24 +405,115 @@ namespace Dfe.Academies.Academisation.Domain.UnitTest.SignificantChange
 		[InlineData("2024-07-01", null, SignificantChangeTaskStatus.InProgress)]
 		[InlineData(null, "2024-07-01", SignificantChangeTaskStatus.InProgress)]
 		[InlineData("2024-07-01", "2024-07-02", SignificantChangeTaskStatus.Completed)]
-        public void GetProjectDates_ShouldHaveCorrectStatus(string? proposedDecisionDateString, string? proposedChangeDateString, SignificantChangeTaskStatus expectedTaskStatus)
-        {
-            var project = SignificantChangeProject.Create(
-            _fixture.Create<int>(),
-            _fixture.Create<byte>(),
-            _fixture.Create<string>(),
-            _fixture.Create<string>(),
-            _fixture.Create<string>(),
-            _fixture.Create<string>(),
-            DateTime.UtcNow);
+		public void GetProjectDates_ShouldHaveCorrectStatus(string? proposedDecisionDateString,
+			string? proposedChangeDateString, SignificantChangeTaskStatus expectedTaskStatus)
+		{
+			var project = SignificantChangeProject.Create(
+				new SignificantChangeProjectOptions(
+					_fixture.Create<int>(),
+					_fixture.Create<byte>(),
+					_fixture.Create<string>(),
+					_fixture.Create<string>(),
+					_fixture.Create<string>(),
+					_fixture.Create<string>()),
+				DateTime.UtcNow);
 
-            DateTime? proposedDecisionDate = string.IsNullOrEmpty(proposedDecisionDateString) ? null : DateTime.Parse(proposedDecisionDateString);
-            DateTime? proposedChangeDate = string.IsNullOrEmpty(proposedChangeDateString) ? null : DateTime.Parse(proposedChangeDateString);
+			DateTime? proposedDecisionDate = string.IsNullOrEmpty(proposedDecisionDateString)
+				? null
+				: DateTime.Parse(proposedDecisionDateString);
+			DateTime? proposedChangeDate = string.IsNullOrEmpty(proposedChangeDateString)
+				? null
+				: DateTime.Parse(proposedChangeDateString);
 
-            project.Details.ProposedDecisionDate = proposedDecisionDate;
-            project.Details.ProposedChangeDate = proposedChangeDate;
+			project.Details.ProposedDecisionDate = proposedDecisionDate;
+			project.Details.ProposedChangeDate = proposedChangeDate;
 
-            project.Details.GetConfirmProjectDatesTaskStatus().Should().Be(expectedTaskStatus);
-        }
+			project.Details.GetConfirmProjectDatesTaskStatus().Should().Be(expectedTaskStatus);
+		}
+
+		[Fact]
+		public void GetEqualitiesTaskStatus_WhenNoValues_ReturnNotStarted()
+		{
+			var project = SignificantChangeProject.Create(
+				new SignificantChangeProjectOptions(
+					_fixture.Create<int>(),
+					_fixture.Create<byte>(),
+					_fixture.Create<string>(),
+					_fixture.Create<string>(),
+					_fixture.Create<string>(),
+					_fixture.Create<string>()),
+				DateTime.UtcNow);
+
+
+			project.Details.GetEqualitiesTaskStatus().Should().Be(SignificantChangeTaskStatus.NotStarted);
+		}
+
+		[Theory]
+		[InlineData(true)]
+		[InlineData(false)]
+		public void GetEqualitiesTaskStatus_WhenSetAssessmentHasBeenCompleted_ReturnInprogress(
+			bool equalitiesImpactAssessmentCompleted)
+		{
+			var project = SignificantChangeProject.Create(
+				new SignificantChangeProjectOptions(
+					_fixture.Create<int>(),
+					_fixture.Create<byte>(),
+					_fixture.Create<string>(),
+					_fixture.Create<string>(),
+					_fixture.Create<string>(),
+					_fixture.Create<string>()),
+				DateTime.UtcNow);
+
+
+			project.SetEqualitiesImpactAssessment(equalitiesImpactAssessmentCompleted, null, null);
+
+			project.Details.GetEqualitiesTaskStatus().Should().Be(SignificantChangeTaskStatus.InProgress);
+		}
+
+		[Theory]
+		[InlineData(EqualitiesImpact.None)]
+		[InlineData(EqualitiesImpact.PotentialImpacts)]
+		public void GetEqualitiesTaskStatus_WhenImpactsHaveBeenSet_ReturnCompleted(EqualitiesImpact equalitiesImpact)
+		{
+			var project = SignificantChangeProject.Create(
+				new SignificantChangeProjectOptions(
+					_fixture.Create<int>(),
+					_fixture.Create<byte>(),
+					_fixture.Create<string>(),
+					_fixture.Create<string>(),
+					_fixture.Create<string>(),
+					_fixture.Create<string>()),
+				DateTime.UtcNow);
+
+
+			project.SetEqualitiesImpactAssessment(true, equalitiesImpact, null);
+
+			project.Details.GetEqualitiesTaskStatus().Should().Be(SignificantChangeTaskStatus.Completed);
+		}
+
+		[Theory]
+		[InlineData(EqualitiesImpact.ImpactsIdentified, "", SignificantChangeTaskStatus.Completed)]
+		[InlineData(EqualitiesImpact.ImpactsIdentified, null, SignificantChangeTaskStatus.Completed)]
+		[InlineData(EqualitiesImpact.ImpactsIdentified, "Mitigation plan in place",
+			SignificantChangeTaskStatus.Completed)]
+		public void GetEqualitiesTaskStatus_WhenImpactsHaveBeenIdentified_ShouldReturnCorrectStatus(
+			EqualitiesImpact impact, string? mitigation, SignificantChangeTaskStatus expectedStatus)
+		{
+			var project = SignificantChangeProject.Create(
+				new SignificantChangeProjectOptions(
+					_fixture.Create<int>(),
+					_fixture.Create<byte>(),
+					_fixture.Create<string>(),
+					_fixture.Create<string>(),
+					_fixture.Create<string>(),
+					_fixture.Create<string>()),
+				DateTime.UtcNow);
+
+
+			project.SetEqualitiesImpactAssessment(true, impact, mitigation);
+
+			project.Details.GetEqualitiesTaskStatus().Should().Be(expectedStatus);
+		}
 	}
 }
+