+++ Dfe.Academies.Academisation.IService/ServiceModels/SignificantChange/SignificantChangeProjectDto.cs
@@ -12,10 +12,19 @@ public class SignificantChangeProjectDto
 	public string? AssignedUserFullName { get; set; }
 	public string? AssignedUserEmailAddress { get; set; }
 	public string TypeOfSignificantChange { get; set; } = string.Empty;
+	public string? LocalAuthorityName { get; set; }
+	public string? CompaniesHouseNumber { get; set; }
 	public string Status { get; set; } = string.Empty;
 	public bool? TrustConsultedStakeholders { get; set; }
 	public string? TrustConsultedStakeholdersNotConsultedReason { get; set; }
 	public string StakeholderConsultationTaskStatus { get; set; } = string.Empty;
+	public string EqualitiesTaskStatus { get; set; } = string.Empty;
+	public bool? EqualitiesImpactAssessmentCompleted { get; set; }
+	public string? EqualitiesImpactIdentified { get; set; }
+	public string? EqualitiesImpactIdentifiedMitigation { get; set; }
+	public bool? TrustConsultedReligiousBody { get; set; }
+	public string? TrustConsultedReligiousBodyNotConsultedReason { get; set; }
+	public string ReligiousBodyConsultationTaskStatus { get; set; } = string.Empty;
 	public DateTime? ProposedDecisionDate { get; set; }
 	public DateTime? ProposedChangeDate { get; set; }
 	public string ConfirmProjectDatesTaskStatus { get; set; } = string.Empty;