+++ Dfe.Academies.Academisation.Service.UnitTest/Queries/SignificantChange/GetSignificantChangeProjectByIdQueryHandlerTests.cs
@@ -37,18 +37,23 @@ namespace Dfe.Academies.Academisation.Service.UnitTest.Queries.SignificantChange
 
 			var project = new SignificantChangeProject(
 				SignificantChangeStatus.PreDecision,
-				urn: 123456,
-				tier: 2,
-				trustName: "Trust A",
-				trustUkprn: "10000001",
-				typeOfSignificantChange: "Change of age range",
-				schoolName: "School A")
+				new SignificantChangeProjectOptions(
+					urn: 123456,
+					tier: 2,
+					trustName: "Trust A",
+					trustUkprn: "10000001",
+					typeOfSignificantChange: "Change of age range",
+					schoolName: "School A"
+				)
+			)
 			{
 				Id = query.Id
 			};
 
 			project.AssignUser(assignedUserId, "assigned.user@test.local", "Assigned User");
 			project.SetStakeholderConsultation(false, "Trust has not consulted stakeholders yet");
+			project.SetEqualitiesImpactAssessment(true, EqualitiesImpact.ImpactsIdentified, "Mitigation");
+			project.SetReligiousBodyConsultation(false, "Trust has not consulted religious body yet");
 			project.SetProjectDates(proposedDecisionDate, proposedChangeDate);
 
 			_repositoryMock
@@ -70,6 +75,14 @@ namespace Dfe.Academies.Academisation.Service.UnitTest.Queries.SignificantChange
 			result.StakeholderConsultation.TrustConsultedStakeholders.Should().BeFalse();
 			result.StakeholderConsultation.TrustConsultedStakeholdersNotConsultedReason.Should().Be("Trust has not consulted stakeholders yet");
 			result.StakeholderConsultation.Status.Should().Be(nameof(SignificantChangeTaskStatus.Completed));
+
+			result.EqualitiesImpactAssessment.EqualitiesImpactAssessmentCompleted.Should().BeTrue();
+			result.EqualitiesImpactAssessment.EqualitiesImpactIdentified.Should().Be(nameof(EqualitiesImpact.ImpactsIdentified));
+			result.EqualitiesImpactAssessment.EqualitiesImpactIdentifiedMitigation.Should().Be("Mitigation");
+			result.EqualitiesImpactAssessment.Status.Should().Be(nameof(SignificantChangeTaskStatus.Completed));
+			result.ReligiousBodyConsultation.TrustConsultedReligiousBody.Should().BeFalse();
+			result.ReligiousBodyConsultation.TrustConsultedReligiousBodyNotConsultedReason.Should().Be("Trust has not consulted religious body yet");
+			result.ReligiousBodyConsultation.Status.Should().Be(nameof(SignificantChangeTaskStatus.Completed));
 			result.ProjectDates.ProposedDecisionDate.Should().Be(proposedDecisionDate);
 			result.ProjectDates.ProposedChangeDate.Should().Be(proposedChangeDate);
 			result.ProjectDates.Status.Should().Be(nameof(SignificantChangeTaskStatus.Completed));