+++ Dfe.Academies.Academisation.Service.UnitTest/Queries/SignificantChange/GetSignificantProjectsQueryHandlerTests.cs
@@ -45,6 +45,8 @@ namespace Dfe.Academies.Academisation.Service.UnitTest.Queries.SignificantChange
 
 			projects[0].AssignUser(assignedUserId, "assigned.user@test.local", "Assigned User");
 			projects[0].SetStakeholderConsultation(false, "Trust has not consulted stakeholders yet");
+			projects[0].SetEqualitiesImpactAssessment(true, EqualitiesImpact.ImpactsIdentified, "Needs mitigating actions");
+			projects[0].SetReligiousBodyConsultation(false, "Trust has not consulted religious body yet");
 			projects[0].SetProjectDates(proposedDecisionDate, proposedChangeDate);
 
 			_repositoryMock
@@ -73,9 +75,17 @@ namespace Dfe.Academies.Academisation.Service.UnitTest.Queries.SignificantChange
 			data[0].StakeholderConsultation.TrustConsultedStakeholders.Should().BeFalse();
 			data[0].StakeholderConsultation.TrustConsultedStakeholdersNotConsultedReason.Should().Be("Trust has not consulted stakeholders yet");
 			data[0].StakeholderConsultation.Status.Should().Be(nameof(SignificantChangeTaskStatus.Completed));
+			data[0].ReligiousBodyConsultation.TrustConsultedReligiousBody.Should().BeFalse();
+			data[0].ReligiousBodyConsultation.TrustConsultedReligiousBodyNotConsultedReason.Should().Be("Trust has not consulted religious body yet");
+			data[0].ReligiousBodyConsultation.Status.Should().Be(nameof(SignificantChangeTaskStatus.Completed));
 			data[0].ProjectDates.ProposedDecisionDate.Should().Be(proposedDecisionDate);
 			data[0].ProjectDates.ProposedChangeDate.Should().Be(proposedChangeDate);
 			data[0].ProjectDates.Status.Should().Be(nameof(SignificantChangeTaskStatus.Completed));
+
+			data[0].EqualitiesImpactAssessment.EqualitiesImpactAssessmentCompleted.Should().BeTrue();
+			data[0].EqualitiesImpactAssessment.EqualitiesImpactIdentified.Should().Be(nameof(EqualitiesImpact.ImpactsIdentified));
+			data[0].EqualitiesImpactAssessment.EqualitiesImpactIdentifiedMitigation.Should().Be("Needs mitigating actions");
+			data[0].EqualitiesImpactAssessment.Status.Should().Be(nameof(SignificantChangeTaskStatus.Completed));
 
 			data[1].Id.Should().Be(11);
 			data[1].Urn.Should().Be(654321);
@@ -87,6 +97,12 @@ namespace Dfe.Academies.Academisation.Service.UnitTest.Queries.SignificantChange
 			data[1].TypeOfSignificantChange.Should().Be("Change of gender composition");
 			data[1].Status.Should().Be(nameof(SignificantChangeStatus.PreDecision));
 			data[1].StakeholderConsultation.Status.Should().Be(nameof(SignificantChangeTaskStatus.NotStarted));
+
+			data[1].EqualitiesImpactAssessment.EqualitiesImpactAssessmentCompleted.Should().BeNull();
+			data[1].EqualitiesImpactAssessment.EqualitiesImpactIdentified.Should().BeNull();
+			data[1].EqualitiesImpactAssessment.EqualitiesImpactIdentifiedMitigation.Should().BeNull();
+			data[1].EqualitiesImpactAssessment.Status.Should().Be(nameof(SignificantChangeTaskStatus.NotStarted));
+			data[1].ReligiousBodyConsultation.Status.Should().Be(nameof(SignificantChangeTaskStatus.NotStarted));
 			data[1].ProjectDates.Status.Should().Be(nameof(SignificantChangeTaskStatus.NotStarted));
 		}
 
@@ -177,7 +193,8 @@ namespace Dfe.Academies.Academisation.Service.UnitTest.Queries.SignificantChange
 
 		private static SignificantChangeProject CreateProject(int id, int urn, byte tier, string trustName, string trustUkprn, string route, string schoolName)
 		{
-			return new SignificantChangeProject(SignificantChangeStatus.PreDecision, urn, tier, trustName, trustUkprn, route, schoolName)
+			return new SignificantChangeProject(SignificantChangeStatus.PreDecision, new SignificantChangeProjectOptions(
+				 urn, tier, trustName, trustUkprn, route, schoolName))
 			{
 				Id = id,
 				CreatedOn = DateTime.UtcNow